# docs/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "SOAT - OS Service",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8081/"
    }
  ],
  "paths": {
...
    "/api/v1/produtos/{id}": {
      "get": {
        "tags": [
          "Produtos"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProdutoOutput"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Produtos"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AtualizarProdutoRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AtualizarProdutoRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AtualizarProdutoRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProdutoOutput"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Produtos"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content",
            "content": {
              "application/json": { }
            }
          }
        }
      }
    },
    "/api/v1/produtos": {
      "get": {
        "tags": [
          "Produtos"
        ],
        "parameters": [
          {
            "name": "Pagina",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "Tamanho",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResultOfProdutoOutput"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Produtos"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InserirProdutoRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/InserirProdutoRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/InserirProdutoRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProdutoOutput"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/produtos/{id}/incrementar-estoque": {
      "patch": {
        "tags": [
          "Produtos"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AtualizarQuantidadeEstoqueProdutoRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AtualizarQuantidadeEstoqueProdutoRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AtualizarQuantidadeEstoqueProdutoRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProdutoOutput"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AtualizarProdutoRequest": {
        "required": [
          "nome",
          "descricao",
          "valor"
        ],
        "type": "object",
        "properties": {
          "nome": {
            "type": "string"
          },
          "descricao": {
            "type": "string"
          },
          "valor": {
            "type": "number",
            "format": "double"
          }
        }
      },
      "AtualizarQuantidadeEstoqueProdutoRequest": {
        "required": [
          "quantidade"
        ],
        "type": "object",
        "properties": {
          "quantidade": {
            "type": "number",
            "format": "double"
          }
        }
      },
...
      "InserirProdutoRequest": {
        "required": [
          "nome",
          "descricao",
          "valor",
          "quantidadeEmEstoque"
        ],
        "type": "object",
        "properties": {
          "nome": {
            "type": "string"
          },
          "descricao": {
            "type": "string"
          },
          "valor": {
            "type": "number",
            "format": "double"
          },
          "quantidadeEmEstoque": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "PagedResultOfProdutoOutput": {
        "required": [
          "items",
          "totalCount",
          "page",
          "pageSize"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProdutoOutput"
            }
          },
          "totalCount": {
            "type": "integer",
            "format": "int32"
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "hasNext": {
            "type": "boolean"
          },
          "hasPrev": {
            "type": "boolean"
          }
        }
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "ProdutoOutput": {
        "required": [
          "id",
          "nome",
          "descricao",
          "valor",
          "quantidadeEmEstoque"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "nome": {
            "type": "string"
          },
          "descricao": {
            "type": "string"
          },
          "valor": {
            "type": "number",
            "format": "double"
          },
          "quantidadeEmEstoque": {
            "type": "number",
            "format": "double"
          }
        }
      },
...
    }
  },
...
}